% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigEGStar.R
\name{bigEGStar}
\alias{bigEGStar}
\title{Finding big E of g star  (middle of page 6)}
\usage{
bigEGStar(doc1, doc2, omega1, cleaned_data, sigma, tolerance, little_omega)
}
\arguments{
\item{doc1}{The document id of the first document you wish to compare}

\item{doc2}{The document id of the second document you wish to compare}

\item{omega1}{The omega matrix}

\item{cleaned_data}{The dataset you are given with comparisons, cleaned using cleaner}

\item{sigma}{The hyperprior chosen to identify the latent space}

\item{tolerance}{The tolerance to use for Newton Method convergence}

\item{little_omega}{The covariance vector for a theoretical new observation on the dataset}
}
\value{
A double
}
\description{
Function that will find the second term of the s function
}
\author{
Group <\email{group@wustl.edu}
}
