% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{cleaner}
\alias{cleaner}
\title{cleaner}
\usage{
cleaner(dataset)
}
\arguments{
\item{dataset}{input dataset}
}
\value{
output dataset
}
\description{
cleans data
}
\author{
Group <\email{group@wustl.edu}
}
