% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2LogLike.R
\name{d2LogLike}
\alias{d2LogLike}
\title{Second derivative of log-likelihood function (Equation 7)}
\usage{
d2LogLike(g, y)
}
\arguments{
\item{g}{An unknown vector of real numbers}

\item{y}{A vector of outcomes of pairwise comparison of documents}
}
\value{
A diagonal Hessian matrix, with ith element along the diagonal
}
\description{
Returns the second derivative of log-likelihood of each ith element with respect to g
}
\author{
Group <\email{group@wustl.edu}
}
