% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2LogPri.R
\name{d2LogPri}
\alias{d2LogPri}
\title{Second derivative of the log prior function (Equation 5)}
\usage{
d2LogPri(omega, g)
}
\arguments{
\item{omega}{A matrix which contains information about how a document is compared with all other documents.}

\item{g}{An unknown vector of real numbers}
}
\value{
A matrix that is the second derivative of the log prior function
}
\description{
Returns the negative of the inverse of the omega matrix
}
\author{
Gangyi Sun <\email{gangyi.sun@wustl.edu}
}
