% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLogLike.R
\name{dLogLike}
\alias{dLogLike}
\title{Derivative of log-likelihood function (Equation 6)}
\usage{
dLogLike(g, y)
}
\arguments{
\item{g}{An unknown vector of real numbers}

\item{y}{A vector of outcomes of pairwise comparison of documents}
}
\value{
A vector with ith element of gradient with respect to g
}
\description{
Returns the derivative of log-likelihood of each ith element with respect to g
}
\author{
Group <\email{group@wustl.edu}
}
