% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLogPri.R
\name{dLogPri}
\alias{dLogPri}
\title{Derivative of the log prior function (Equation 4)}
\usage{
dLogPri(omega, g)
}
\arguments{
\item{omega}{A matrix which contains information about how a document is compared with all other documents.}

\item{g}{An unknown vector of real numbers}
}
\value{
A vector that is the derivative of the log prior function
}
\description{
Returns the product of the negative of the inverse of the omega matrix multiplied by the g vector
}
\author{
Group <\email{group@wustl.edu}
}
