% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalS.R
\name{finalS}
\alias{finalS}
\title{Finding final s (bottom of page 5)
Function that will the s function}
\usage{
finalS(doc1, doc2, dataset, g, sigma, tolerance)
}
\arguments{
\item{doc1}{The document id of the first document you wish to compare}

\item{doc2}{The document id of the second document you wish to compare}

\item{dataset}{The dataset you are given with comparisons}

\item{g}{An unknown vector of real numbers}

\item{sigma}{The hyperprior chosen to identify the latent space}

\item{tolerance}{The tolerance to use for Newton Method convergence}
}
\value{
A double
}
\description{
Finding final s (bottom of page 5)
Function that will the s function
}
\author{
Group <\email{group@wustl.edu}
}
