% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPrior.R
\name{gPrior}
\alias{gPrior}
\title{Finding prior of g (equation 1)}
\usage{
gPrior(omega)
}
\arguments{
\item{omega}{the omega matrix}
}
\value{
A vector, the prior on g
}
\description{
Function that will calculate the prior on g
}
\author{
Group <\email{group@wustl.edu}
}
