% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLike.R
\name{logLike}
\alias{logLike}
\title{Log-likelihood function (Equation 5.9)}
\usage{
logLike(g, y)
}
\arguments{
\item{g}{An unknown vector of real numbers}

\item{y}{A vector of outcomes of pairwise comparison of documents}
}
\value{
A numeric which is the sum of log-likelihoods of all comparisons
}
\description{
Returns the log-likelihood given a set of observed preferences y and the difference function g
}
\author{
Group <\email{group@wustl.edu}
}
