% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowercaseH.R
\name{lowercaseH}
\alias{lowercaseH}
\title{lowercase H function at the top of page 6}
\usage{
lowercaseH(x)
}
\arguments{
\item{x}{a number between 0 and 1}
}
\value{
A double
}
\description{
Function to be used to calculate capital H
}
\author{
Group <\email{group@wustl.edu}
}
