% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxInfoComp.R
\name{maxInfoComp}
\alias{maxInfoComp}
\title{Finding maximum information comparison}
\usage{
maxInfoComp(dataset, sigma, tolerance, nComp)
}
\arguments{
\item{dataset}{The dataset you are given with comparisons}

\item{sigma}{The hyperprior chosen to identify the latent space}

\item{tolerance}{The tolerance to use for Newton Method convergence}
}
\value{
A vector with two entries containing document ids for your new comparison
}
\description{
Function that will calculate the comparison that gives the most information
}
\author{
Group <\email{group@wustl.edu}
}
