% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muMaker.R
\name{muMaker}
\alias{muMaker}
\title{Finding mu hat for equation 14 (bottom of page 4)}
\usage{
muMaker(doc1, doc2, omega1, cleaned_data, sigma, tolerance, little_omega)
}
\arguments{
\item{doc1}{The document id of the first document you wish to compare}

\item{doc2}{The document id of the second document you wish to compare}

\item{omega1}{The omega matrix}

\item{cleaned_data}{The dataset you are given with comparisons, cleaned using cleaner}

\item{sigma}{The hyperprior chosen to identify the latent space}

\item{tolerance}{The tolerance to use for Newton Method convergence}

\item{little_omega}{The covariance vector for a theoretical new observation on the dataset}
}
\value{
A double
}
\description{
Function that will find expected value of g for a new comparison
}
\author{
Group <\email{group@wustl.edu}
}
