% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtonMethod.R
\name{newtonMethod}
\alias{newtonMethod}
\title{Newton appriximation function (Equation 9.5)}
\usage{
newtonMethod(omega, y, tolerance)
}
\arguments{
\item{omega}{A matrix which contains information about how a document is compared with all other documents.}

\item{y}{A vector of outcomes of pairwise comparison of documents}

\item{tolerance}{A given value of tolerance}

\item{g}{An unknown vector of real numbers}
}
\value{
A vector gHat which is the approximate distribution
}
\description{
Calculating the approximate distribution using Newton's method to iteratively to find gHat
}
\author{
Group <\email{group@wustl.edu}
}
