% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtonMethodPrequel.R
\name{newtonMethodPrequel}
\alias{newtonMethodPrequel}
\title{Newton method prequel, called in newtonMethod (Equation 9.5)}
\usage{
newtonMethodPrequel(omega, g, y)
}
\arguments{
\item{omega}{A matrix which contains information about how a document is compared with all other documents.}

\item{g}{An unknown vector of real numbers}

\item{y}{A vector of outcomes of pairwise comparison of documents}
}
\value{
A vector
}
\description{
Function needed for the Newton Approximation method
}
\author{
Group <\email{group@wustl.edu}
}
