% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega.R
\name{omega}
\alias{omega}
\title{Calculating the omega matrix (Equation 1.5)}
\usage{
omega(dataset, sigma)
}
\arguments{
\item{dataset}{A dataframe of comparisons, with each document being identified by an unique id.}

\item{sigma}{A numeric value of sigma, which is a hyper-prior chosen to identify the latent space.}
}
\value{
A matrix which contains information about how a document is compared with all other documents.
}
\description{
Returns the omega matrix given an input of pair-wise comparisons and a sigma value
}
\author{
Gangyi Sun <\email{gangyi.sun@wustl.edu}
}
